refactor: simplefy the folder structure for Home, rename as global.css
