+++ index.html
@@ -48,7 +48,7 @@
 
       gtag("config", "G-4PJ31TXM54");
     </script>
-    <link href="/src/index.css" rel="stylesheet">
+    <link href="/src/global.css" rel="stylesheet">
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
